pyproject.toml
[tool.poetry]
name = "structured_streaming"
version = "0.1.0"
description = ""
authors = ["Alex Vakhmenin <alex@vakhmenin.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyspark = "^3.1.2"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.5b2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
